deps: bump asto to v1.8.0, extract benchmarks as a module
