+++ benchmarks/LICENSE.header
@@ -1,0 +1,2 @@
+The MIT License (MIT) Copyright (c) 2020-2021 artipie.com
+https://github.com/artipie/rpm-adapter/LICENSE.txt
