+++ pom.xml
@@ -111,7 +111,7 @@ SOFTWARE.
     <dependency>
       <groupId>com.artipie</groupId>
       <artifactId>asto</artifactId>
-      <version>v1.7.0</version>
+      <version>v1.8.0</version>
     </dependency>
     <dependency>
       <groupId>commons-cli</groupId>
